everything now scales correctly when resolution changes, hopefully. anything in 16:9 should work. Implemented all of options and background transparency correctly updates now. Menu buttons are now not hard-coded coordinates and are based on the font so they will work regardless of font now. Exit Game and Return now do what they are supposed to in Options screen.

+++ src/com/jwhi/interactivefiction/game/Loading.java
@@ -28,7 +28,7 @@ public class Loading extends BasicGameState {
 	// Draws the loading screen
 	public void render(GameContainer gc, StateBasedGame sbg, Graphics g) throws SlickException {
 //		g.drawImage(loadingImage, 0, 0);
-		loadingImage.draw(0, 0, gc.getWidth()/Constants.MaxResourceWidth);
+		loadingImage.draw(0, 0, (float)gc.getWidth()/(float)Constants.MaxResourceWidth);
 		
 	}
